genai-function-calling: Adds MCP example to Spring AI (#61)
This implements the same flow described in the PR that [adds MCP example
to Vercel](https://github.com/elastic/observability-examples/pull/58).

Traces work as expected (broken when MCP [until the protocol is
fixed](https://github.com/modelcontextprotocol/modelcontextprotocol/issues/246#issuecomment-2796802224)).

Normal is same as before:
![Screenshot 2025-04-15 at 2 15
25 PM](https://github.com/user-attachments/assets/0c44ace5-1b2b-43cc-bae1-f104a0a928d5)

MCP (`--mcp`) launches a sub-process and while that is traced (as it
inherits the agent), the propagation is broken at the moment until spec
support is added.
![Screenshot 2025-04-15 at 2 16
32 PM](https://github.com/user-attachments/assets/f3250b00-b84c-442e-813b-3c113125d4f4)
![Screenshot 2025-04-15 at 2 16
42 PM](https://github.com/user-attachments/assets/98dc079f-f9f8-4660-a4dd-deb1dbc36286)
![Screenshot 2025-04-15 at 2 16
22 PM](https://github.com/user-attachments/assets/71109305-3f9f-487f-a50b-eb40bae7d72b)

---------

Signed-off-by: Adrian Cole <adrian.cole@elastic.co>