genai-function-calling: update to qwen3 model which reduces edge cases (#71)
This updates to [qwen3](https://github.com/QwenLM/Qwen3) model which
reduces edge cases as all except spring-ai can use qwen3:0.6b regardless
of if the `major_version` is needed by the LLM when answering the
question.

Notes:

* a follow-up PR should probably happen to remove the think tags from
reasoning models (qwen3 is reasoning) Here's an example
https://github.com/vercel/ai/blob/main/content/docs/03-ai-sdk-core/40-middleware.mdx#extract-reasoning
Another option is to disable thinking mode, but that's not [yet
implemented in ollama](https://github.com/ollama/ollama/issues/10473).
* That leaving out `major_version` requires a larger model in spring-ai
is likely due to how the tool requests are translated in to a chat
completion schema. I've not looked carefully at why and it is possible a
tweak could be made to spring-ai to prevent needing a larger model here.
This edge case is not new and also qwen2.5:3b was needed before while
now we can get by with qwen3:1.7b

Signed-off-by: Adrian Cole <adrian.cole@elastic.co>